# docs/swagger.yaml
definitions:
  rpc.ApplicationCallback:
    properties:
      application_id:
        type: string
      callback_event:
        type: string
      callback_url:
        type: string
      id:
        type: integer
    type: object
  rpc.ChargePoint:
    properties:
      application_id:
        type: string
      charge_box_serial_number:
        type: string
      charge_point_identifier:
        type: string
      charge_point_model:
        type: string
      charge_point_serial_number:
        type: string
      charge_point_vendor:
        type: string
      connector_count:
        type: integer
      entity_code:
        type: string
      firmware_version:
        type: string
      iccid:
        type: string
      id:
        type: integer
      imsi:
        type: string
      meter_serial_number:
        type: string
      meter_type:
        type: string
      ocpp_protocol:
        type: string
    type: object
  rpc.CreateApplicationCallbackReqPublic:
    properties:
      callback_event:
        type: string
      callback_url:
        type: string
    type: object
  rpc.CreateApplicationCallbackResp:
    properties:
      application_callback:
        $ref: '#/definitions/rpc.ApplicationCallback'
    type: object
  rpc.CreateChargePointIdTagReq:
    properties:
      application_id:
        type: string
      charge_point_identifier:
        type: string
      id_tag:
        type: string
    type: object
  rpc.CreateChargePointIdTagReqPublic:
    properties:
      charge_point_identifier:
        type: string
      id_tag:
        type: string
    type: object
  rpc.CreateChargePointReqPublic:
    properties:
      charge_point_identifier:
        type: string
      entity_code:
        type: string
      ocpp_protocol:
        type: string
    type: object
  rpc.CreateChargePointResp:
    properties:
      charge_point:
        $ref: '#/definitions/rpc.ChargePoint'
    type: object
  rpc.GetApplicationCallbacksResp:
    properties:
      application_callbacks:
        items:
          $ref: '#/definitions/rpc.ApplicationCallback'
        type: array
    type: object
  rpc.GetLatestStatusNotificationsReqPublic:
    properties:
      charge_point_identifier:
        type: string
    type: object
  rpc.GetLatestStatusNotificationsResp:
    properties:
      connector_status:
        items:
          $ref: '#/definitions/rpc.StatusNotification'
        type: array
    type: object
  rpc.RemoteStartTransactionReqPublic:
    properties:
      charge_point_identifier:
        type: string
      connector_id:
        type: integer
    type: object
  rpc.RemoteStartTransactionResp:
    properties:
      transaction_id:
        type: integer
    type: object
  rpc.RemoteStopTransactionReqPublic:
    properties:
      charge_point_identifier:
        type: string
      connector_id:
        type: integer
    type: object
  rpc.RemoteStopTransactionResp:
    type: object
  rpc.StatusNotification:
    properties:
      charge_point_id:
        type: integer
      connector_id:
        type: integer
      error_code:
        type: string
      id:
        type: integer
      info:
        type: string
      reported_timestamp:
        type: string
      status:
        type: string
      timestamp:
        type: string
      vendor_error_code:
        type: string
      vendor_id:
        type: string
    type: object
host: ocpp-dev.beepbeep.tech:8060
info:
  contact:
    email: lowen@beepbeep.tech
    name: Lowen
  description: Service to interface with OCPP-compliant charge points
  title: BB3 OCPP API
  version: "2.0"
paths:
  /v2/ocpp/applications/callbacks:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rpc.GetApplicationCallbacksResp'
      security:
      - ApiKeyAuth: []
      summary: Get Callback URL for Application
      tags:
      - Applications
    post:
      consumes:
      - application/json
      parameters:
      - description: Post CreateApplicationCallbackReq body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/rpc.CreateApplicationCallbackReqPublic'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rpc.CreateApplicationCallbackResp'
      security:
      - ApiKeyAuth: []
      summary: Set Callback URL for Application
      tags:
      - Applications
  /v2/ocpp/charge_points:
    post:
      consumes:
      - application/json
      parameters:
      - description: Post CreateChargePointReq body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/rpc.CreateChargePointReqPublic'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rpc.CreateChargePointResp'
      security:
      - ApiKeyAuth: []
      summary: Create a Charge Point
      tags:
      - Charge Points
  /v2/ocpp/charge_points/id_tags:
    post:
      consumes:
      - application/json
      parameters:
      - description: Post CreateChargePointIdTagReq body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/rpc.CreateChargePointIdTagReqPublic'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rpc.CreateChargePointIdTagReq'
      security:
      - ApiKeyAuth: []
      summary: Create a Charge Point ID tag
      tags:
      - Charge Points
  /v2/ocpp/operations/get-latest-status:
    post:
      consumes:
      - application/json
      parameters:
      - description: Post GetLatestStatus body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/rpc.GetLatestStatusNotificationsReqPublic'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rpc.GetLatestStatusNotificationsResp'
      security:
      - ApiKeyAuth: []
      summary: Get Status of Connectors
      tags:
      - Operations
  /v2/ocpp/operations/remote-start-transaction:
    post:
      consumes:
      - application/json
      parameters:
      - description: Post RemoteStartTransactionReq body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/rpc.RemoteStartTransactionReqPublic'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rpc.RemoteStartTransactionResp'
      security:
      - ApiKeyAuth: []
      summary: Request Charge Point to Start a Transaction
      tags:
      - Operations
  /v2/ocpp/operations/remote-stop-transaction:
    post:
      consumes:
      - application/json
      parameters:
      - description: Post RemoteStopTransactionReq body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/rpc.RemoteStopTransactionReqPublic'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rpc.RemoteStopTransactionResp'
      security:
      - ApiKeyAuth: []
      summary: Request Charge Point to Stop a Transaction
      tags:
      - Operations
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"